+++ modules/apps/portal-instances/portal-instances-api/bnd.bnd
@@ -2,9 +2,9 @@ Bundle-Name: Liferay Portal Instances API
 Bundle-SymbolicName: com.liferay.portal.instances.api
 Bundle-Version: 12.0.1
 Export-Package:\
-	com.liferay.portal.instances.adder,\
 	com.liferay.portal.instances.background.task,\
 	com.liferay.portal.instances.background.task.constants,\
 	com.liferay.portal.instances.constants,\
 	com.liferay.portal.instances.exception,\
-	com.liferay.portal.instances.exporter
+	com.liferay.portal.instances.exporter,\
+	com.liferay.portal.instances.manager
