+++ modules/apps/portal-instances/portal-instances-api/src/main/java/com/liferay/portal/instances/manager/PortalInstanceManager.java
@@ -3,17 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.instances.adder;
+package com.liferay.portal.instances.manager;
 
 import com.liferay.portal.kernel.exception.PortalException;
-
-import org.osgi.annotation.versioning.ProviderType;
 
 /**
  * @author Luis Ortiz
  */
-@ProviderType
-public interface PortalInstanceAdder {
+public interface PortalInstanceManager {
 
 	public long addPortalInstance(
 			long userId, String webId, String virtualHostname, String mx,
