+++ modules/apps/portal-instances/portal-instances-api/src/main/resources/com/liferay/portal/instances/manager/packageinfo
